Provisioning: Split ResourcesToMigrate into focused modules and clarify copy
- Extract the folder-row component into FolderEntry.tsx and the sort helper
  (SortKey + compareFolders) into sorting.ts (+ sorting.test.ts), leaving
  ResourcesToMigrate.tsx as just the panel.
- Reword the subfolder-only hint to be clearer: "No resources directly here —
  they're in subfolders. Selecting this folder migrates everything inside it."
