+++ public/app/features/provisioning/Migrate/ResourcesToMigrate.tsx
@@ -1,48 +1,14 @@
-import { css, cx } from '@emotion/css';
+import { css } from '@emotion/css';
 import { useMemo, useState } from 'react';
 
 import { type GrafanaTheme2 } from '@grafana/data';
 import { t, Trans } from '@grafana/i18n';
-import {
-  Button,
-  Checkbox,
-  Combobox,
-  EmptyState,
-  FilterInput,
-  Icon,
-  IconButton,
-  Stack,
-  Text,
-  useStyles2,
-} from '@grafana/ui';
+import { Button, Checkbox, Combobox, EmptyState, FilterInput, Stack, Text, useStyles2 } from '@grafana/ui';
 
+import { FolderEntry } from './FolderEntry';
 import { type FolderRow } from './hooks/useFolderMigrationData';
 import { isMigratableFolder } from './selection';
-
-export type SortKey = 'count-desc' | 'count-asc' | 'title-asc' | 'title-desc';
-
-export function compareFolders(a: FolderRow, b: FolderRow, key: SortKey): number {
-  switch (key) {
-    case 'count-desc':
-      if (b.dashboardCount !== a.dashboardCount) {
-        return b.dashboardCount - a.dashboardCount;
-      }
-      // eslint-disable-next-line @grafana/no-locale-compare
-      return a.title.localeCompare(b.title);
-    case 'count-asc':
-      if (a.dashboardCount !== b.dashboardCount) {
-        return a.dashboardCount - b.dashboardCount;
-      }
-      // eslint-disable-next-line @grafana/no-locale-compare
-      return a.title.localeCompare(b.title);
-    case 'title-asc':
-      // eslint-disable-next-line @grafana/no-locale-compare
-      return a.title.localeCompare(b.title);
-    case 'title-desc':
-      // eslint-disable-next-line @grafana/no-locale-compare
-      return b.title.localeCompare(a.title);
-  }
-}
+import { type SortKey, compareFolders } from './sorting';
 
 interface Props {
   folders: FolderRow[];
@@ -266,91 +232,6 @@ export function ResourcesToMigrate({
   );
 }
 
-interface FolderEntryProps {
-  folder: FolderRow;
-  isExpanded: boolean;
-  isSelected: boolean;
-  selectedDashboardUids: Set<string>;
-  folderCoveredDashboardUids: Set<string>;
-  onToggleExpanded: () => void;
-  onToggleFolder: () => void;
-  onToggleDashboard: (uid: string) => void;
-}
-
-function FolderEntry({
-  folder,
-  isExpanded,
-  isSelected,
-  selectedDashboardUids,
-  folderCoveredDashboardUids,
-  onToggleExpanded,
-  onToggleFolder,
-  onToggleDashboard,
-}: FolderEntryProps) {
-  const styles = useStyles2(getStyles);
-  return (
-    <div className={cx(styles.row, isSelected && styles.rowSelected)}>
-      <div className={styles.rowHeader}>
-        <IconButton
-          name={isExpanded ? 'angle-down' : 'angle-right'}
-          aria-label={
-            isExpanded
-              ? t('provisioning.migrate.resources-collapse', 'Collapse {{folder}}', { folder: folder.title })
-              : t('provisioning.migrate.resources-expand', 'Expand {{folder}}', { folder: folder.title })
-          }
-          onClick={onToggleExpanded}
-        />
-        <Checkbox
-          value={isSelected}
-          onChange={onToggleFolder}
-          aria-label={t('provisioning.migrate.resources-select-folder', 'Select folder {{folder}}', {
-            folder: folder.title,
-          })}
-        />
-        <Icon name="folder" />
-        <Stack direction="column" gap={0} flex={1}>
-          <Text>{folder.title}</Text>
-          <Text variant="bodySmall" color="secondary">
-            {t('provisioning.migrate.resources-folder-summary', '', {
-              count: folder.dashboardCount,
-              defaultValue_one: '{{count}} resource',
-              defaultValue_other: '{{count}} resources',
-            })}
-          </Text>
-        </Stack>
-      </div>
-      {isExpanded && (
-        <div className={styles.children}>
-          {folder.directDashboards.length === 0 ? (
-            <Text variant="bodySmall" color="secondary">
-              <Trans i18nKey="provisioning.migrate.resources-folder-only-subfolders">
-                Resources in this folder live in subfolders. Migrate the folder to bring them all in one go.
-              </Trans>
-            </Text>
-          ) : (
-            folder.directDashboards.map((dash) => {
-              const coveredByFolder = folderCoveredDashboardUids.has(dash.uid);
-              const checked = coveredByFolder || selectedDashboardUids.has(dash.uid);
-              return (
-                <div key={`dash-${dash.uid}`} className={styles.childRow}>
-                  <Checkbox
-                    value={checked}
-                    disabled={coveredByFolder}
-                    onChange={() => onToggleDashboard(dash.uid)}
-                    aria-label={dash.title}
-                  />
-                  <Icon name="apps" size="sm" />
-                  <Text variant="bodySmall">{dash.title}</Text>
-                </div>
-              );
-            })
-          )}
-        </div>
-      )}
-    </div>
-  );
-}
-
 const getStyles = (theme: GrafanaTheme2) => ({
   panel: css({
     display: 'flex',
@@ -378,41 +259,5 @@ const getStyles = (theme: GrafanaTheme2) => ({
     display: 'flex',
     flexDirection: 'column',
     gap: theme.spacing(0.5),
-  }),
-  row: css({
-    display: 'flex',
-    flexDirection: 'column',
-    padding: theme.spacing(1, 1.25),
-    borderRadius: theme.shape.radius.default,
-    border: `1px solid ${theme.colors.border.weak}`,
-    background: theme.colors.background.primary,
-  }),
-  rowSelected: css({
-    borderColor: theme.colors.primary.border,
-    background: theme.colors.background.secondary,
-  }),
-  rowHeader: css({
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-  }),
-  children: css({
-    display: 'flex',
-    flexDirection: 'column',
-    gap: theme.spacing(0.25),
-    paddingLeft: theme.spacing(5),
-    paddingTop: theme.spacing(0.75),
-  }),
-  childRow: css({
-    display: 'flex',
-    flexDirection: 'row',
-    alignItems: 'center',
-    gap: theme.spacing(1),
-    padding: theme.spacing(0.5, 1),
-    borderRadius: theme.shape.radius.default,
-    '&:hover': {
-      background: theme.colors.background.canvas,
-    },
   }),
 });
